Updated Plots for power, power with neighbor, performance profile, p-value of real data
Folder '1' is for CorrIndTest, including 20 simulation types for
increasing sample size at fixed dimension m. It compares rank dcorr with
knn, dcorr, mdcorr, and HHG. It also plots the performance profile at
n=40 for all simulations (m is different though), clearly rdcorr is the
best.

Folder '1 with neighbor' is based on the same results, but plots the
power with respect to increasing neighborhood at fixed n and fixed
dimension. We observe that for linear datasets the optimal k=n-1, and
for nonlinear datasets the optimal k<n-1.

Folder '2' and '2 with neighbor repeats the same plots', but add knn to
dcorr and mdcorr as well. We can see mdcorr is comparable to rdcorr this
time, both are better than dcorr and HHG. Two performance profiles are
n=40 and n=n_{max}/2 are provided. In this sense, if we cannot work out
the consistency of rdcorr, mdcorr with knn is still a method with
theoretical consistency and huge performance improvement; although the
emphasize of the paper may be different if we chose to go that route.

For RealData1, it includes the p-value for Brain CxP, and Shape vs
Disease. Clearly rdcorr is the superior method.

For RealData2, it considers dcorr and mdcorr with knn again. All three
methods perform quite well now, although rdcorr is still the only method
of significance for brain CxP.